Add Redux for state management and persist the state to local storage
New Redux store has been implemented in the admin-dashboard for state management. This commit includes adding @reduxjs/toolkit for creating the store, actions, and reducers and redux-persist to seamlessly persist part of the redux store to local storage.

Corresponding changes in admin-dashboard's package.json reflect the addition of new libraries, including @reduxjs/toolkit and redux-persist.

Furthermore, a new Dockerfile is added to the auth-microservice and amendments made to its .dockerignore file.

Also, there are minor version updates and removal of commented out code on the mobile side.

These changes allow for better state management in the frontend, ensuring a smoother user experience, and setting the stage for future features that rely on global state.